+++ src/app/(route)/customer/(route)/feedback/feedback-info/[feedbackId]/_components/organisms/FeedbackMeta.tsx
@@ -1,0 +1,66 @@
+"use client";
+
+import NewsDetailGnb from "@/app/(route)/news/_components/newsGnb/NewsDetailGnb";
+import { cn } from "@/utils";
+import React, { useState } from "react";
+import SendCommentBox from "@/app/_components/_comment/SendCommentBox";
+import { CommentItem } from "@/_types/comment";
+import { FeedbackInfoType } from "../../../../_types/FeedbackInfoType";
+import FeedbackDetailContent from "../molecules/FeedbackDetailContent";
+
+// TODO: 리팩터링 테스트, 추가 리팩터링 필요
+
+interface FeedbackMetaProps {
+  feedbackInfoData: FeedbackInfoType;
+  id: string | string[];
+  adminRole: "USER" | "ADMIN" | undefined;
+}
+
+const FeedbackMeta = ({
+  feedbackInfoData,
+  id,
+  adminRole,
+}: FeedbackMetaProps) => {
+  const [parentsComment, setParentsComment] = useState<CommentItem | null>(
+    null
+  );
+
+  return (
+    <>
+      {/* mobile gnb */}
+      <div className={cn("pc:hidden tablet:hidden")}>
+        <NewsDetailGnb
+          title={feedbackInfoData?.title}
+          type="feedback"
+          data={feedbackInfoData}
+          id={Number(id)}
+        />
+      </div>
+
+      {/* detail content */}
+      <FeedbackDetailContent
+        id={id}
+        adminRole={adminRole}
+        feedbackInfoData={feedbackInfoData}
+        setParentsComment={setParentsComment}
+      />
+
+      {/* send comment */}
+      <div
+        className={cn(
+          "shadow-soft-md sticky bottom-0 z-50",
+          "mobile:shadow-none mobile:border-b"
+        )}
+      >
+        <SendCommentBox
+          id={id.toString()}
+          type="IMPROVEMENT"
+          parentsComment={parentsComment}
+          setParentsComment={setParentsComment}
+        />
+      </div>
+    </>
+  );
+};
+
+export default FeedbackMeta;
